+++ flag/service/app/app.go
@@ -1,0 +1,5 @@
+package app
+
+type App struct {
+	Unique string
+}